+++ hdrezka/post/urls/__init__.py
@@ -1,0 +1,3 @@
+from .kind import *
+from .short import *
+from .urls import *