+++ hdrezka/post/urls/kind/__init__.py
@@ -1,0 +1,3 @@
+from .quality import *
+from .subtitle import *
+from .video import *